factor out a package for github specific things
