+++ main.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/codeEmitter/gitrob/core"
+	"github.com/codeEmitter/gitrob/github"
 	"github.com/codeEmitter/gitrob/scm"
 )
 
@@ -20,7 +21,7 @@ func GatherTargets(sess *core.Session) {
 	sess.Stats.Status = core.StatusGathering
 	sess.Out.Important("Gathering targets...\n")
 	for _, login := range sess.Options.Logins {
-		target, err := scm.GetUserOrOrganization(login, sess.GithubClient)
+		target, err := github.GetUserOrOrganization(login, sess.GithubClient)
 		if err != nil {
 			sess.Out.Error(" Error retrieving information on %s: %s\n", login, err)
 			continue
@@ -29,7 +30,7 @@ func GatherTargets(sess *core.Session) {
 		sess.AddTarget(target)
 		if *sess.Options.NoExpandOrgs == false && *target.Type == "Organization" {
 			sess.Out.Debug("Gathering members of %s (ID: %d)...\n", *target.Login, *target.ID)
-			members, err := scm.GetOrganizationMembers(target.Login, sess.GithubClient)
+			members, err := github.GetOrganizationMembers(target.Login, sess.GithubClient)
 			if err != nil {
 				sess.Out.Error(" Error retrieving members of %s: %s\n", *target.Login, err)
 				continue
@@ -63,7 +64,7 @@ func GatherRepositories(sess *core.Session) {
 					wg.Done()
 					return
 				}
-				repos, err := scm.GetRepositoriesFromOwner(target.Login, sess.GithubClient)
+				repos, err := github.GetRepositoriesFromOwner(target.Login, sess.GithubClient)
 				if err != nil {
 					sess.Out.Error(" Failed to retrieve repositories from %s: %s\n", *target.Login, err)
 				}
